dom/websocket/tests/test_webSocket_sandbox.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1252751</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<div id="container"></div>
<iframe id="frame"></iframe>
<script type="application/javascript">
var urls = [ "https://example.com/tests/dom/websocket/tests/iframe_webSocket_sandbox.html",
             "https://example.com/tests/dom/websocket/tests/iframe_webSocket_sandbox.html?nested",
             "https://example.com/tests/dom/websocket/tests/iframe_webSocket_sandbox.html?popup" ];

onmessage = function(e) {
  is(e.data, "WS Throws!", "ws://URI cannot be used by a https iframe");
  runTest();
}

function runTest() {
  if (!urls.length) {
    SimpleTest.finish();
    return;
  }

  document.getElementById("frame").src = urls.shift();
}

SimpleTest.waitForExplicitFinish();
runTest();
</script>
</body>
</html>
